Refactor brave_wallet_provider_delegate_impl to be shared with android
