+++ browser/brave_wallet/brave_wallet_provider_delegate_impl.cc
@@ -10,12 +10,10 @@
 #include <vector>
 
 #include "base/bind.h"
-#include "brave/browser/brave_wallet/brave_wallet_tab_helper.h"
+#include "brave/browser/brave_wallet/brave_wallet_provider_delegate_impl_helper.h"
 #include "brave/browser/brave_wallet/keyring_controller_factory.h"
-#include "brave/browser/ui/brave_pages.h"
 #include "brave/components/brave_wallet/browser/keyring_controller.h"
 #include "brave/components/permissions/contexts/brave_ethereum_permission_context.h"
-#include "chrome/browser/ui/browser_finder.h"
 #include "components/content_settings/core/common/content_settings.h"
 #include "content/public/browser/render_frame_host.h"
 #include "content/public/browser/web_contents.h"
@@ -48,8 +46,6 @@ void OnGetAllowedAccounts(
   std::move(callback).Run(success, allowed_accounts);
 }
 
-base::OnceCallback<void()> g_NewSetupNeededForTestingCallback;
-
 }  // namespace
 
 BraveWalletProviderDelegateImpl::BraveWalletProviderDelegateImpl(
@@ -67,25 +63,13 @@ BraveWalletProviderDelegateImpl::BraveWalletProviderDelegateImpl(
 
 BraveWalletProviderDelegateImpl::~BraveWalletProviderDelegateImpl() = default;
 
-// static
-void BraveWalletProviderDelegateImpl::SetCallbackForNewSetupNeededForTesting(
-    base::OnceCallback<void()> callback) {
-  g_NewSetupNeededForTestingCallback = std::move(callback);
-}
-
 GURL BraveWalletProviderDelegateImpl::GetOrigin() const {
   auto* rfh = content::RenderFrameHost::FromID(host_id_);
   return rfh ? rfh->GetLastCommittedURL().GetOrigin() : GURL();
 }
 
-void BraveWalletProviderDelegateImpl::ShowBubble() {
-  if (!web_contents_)
-    return;
-
-  auto* tab_helper =
-      brave_wallet::BraveWalletTabHelper::FromWebContents(web_contents_);
-  if (tab_helper)
-    tab_helper->ShowBubble();
+void BraveWalletProviderDelegateImpl::ShowPanel() {
+  ::brave_wallet::ShowPanel(web_contents_);
 }
 
 void BraveWalletProviderDelegateImpl::RequestEthereumPermissions(
@@ -122,14 +106,7 @@ void BraveWalletProviderDelegateImpl::
         const std::vector<std::string>& allowed_accounts,
         brave_wallet::mojom::KeyringInfoPtr keyring_info) {
   if (!keyring_info->is_default_keyring_created) {
-    Browser* browser = web_contents_
-                           ? chrome::FindBrowserWithWebContents(web_contents_)
-                           : nullptr;
-    if (browser) {
-      brave::ShowBraveWalletOnboarding(browser);
-    } else if (g_NewSetupNeededForTestingCallback) {
-      std::move(g_NewSetupNeededForTestingCallback).Run();
-    }
+    ShowWalletOnboarding(web_contents_);
     std::move(callback).Run(false, std::vector<std::string>());
     return;
   }
