+++ browser/brave_wallet/brave_wallet_provider_delegate_impl.h
@@ -36,14 +36,11 @@ class BraveWalletProviderDelegateImpl : public BraveWalletProviderDelegate,
       const BraveWalletProviderDelegateImpl&) = delete;
   ~BraveWalletProviderDelegateImpl() override;
 
-  void ShowBubble() override;
+  void ShowPanel() override;
   GURL GetOrigin() const override;
   void RequestEthereumPermissions(
       RequestEthereumPermissionsCallback callback) override;
-
   void GetAllowedAccounts(GetAllowedAccountsCallback callback) override;
-  static void SetCallbackForNewSetupNeededForTesting(
-      base::OnceCallback<void()>);
 
  private:
   void ContinueRequestEthereumPermissions(
