--- browser/brave_wallet/brave_wallet_provider_delegate_impl_android.cc
@@ -1,34 +1,0 @@
-/* Copyright (c) 2021 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#include <string>
-#include <utility>
-#include <vector>
-
-#include "brave/browser/brave_wallet/brave_wallet_provider_delegate_impl_android.h"
-#include "content/public/browser/render_frame_host.h"
-#include "content/public/browser/web_contents.h"
-
-namespace brave_wallet {
-
-BraveWalletProviderDelegateImplAndroid::BraveWalletProviderDelegateImplAndroid(
-    content::WebContents* web_contents,
-    content::RenderFrameHost* const render_frame_host) {}
-
-void BraveWalletProviderDelegateImplAndroid::RequestEthereumPermissions(
-    BraveWalletProviderDelegate::RequestEthereumPermissionsCallback callback) {
-  std::move(callback).Run(false, std::vector<std::string>());
-}
-
-void BraveWalletProviderDelegateImplAndroid::GetAllowedAccounts(
-    BraveWalletProviderDelegate::GetAllowedAccountsCallback callback) {
-  std::move(callback).Run(false, std::vector<std::string>());
-}
-
-GURL BraveWalletProviderDelegateImplAndroid::GetOrigin() const {
-  return GURL();
-}
-
-}  // namespace brave_wallet
