+++ browser/brave_wallet/brave_wallet_provider_delegate_impl_helper_android.cc
@@ -1,0 +1,20 @@
+/* Copyright (c) 2021 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#include "brave/browser/brave_wallet/brave_wallet_provider_delegate_impl_helper.h"
+
+#include "base/notreached.h"
+
+namespace brave_wallet {
+
+void ShowPanel(content::WebContents* web_contents) {
+  NOTIMPLEMENTED();
+}
+
+void ShowWalletOnboarding(content::WebContents* web_contents) {
+  NOTIMPLEMENTED();
+}
+
+}  // namespace brave_wallet
